perf: Improve performance when parsing EMSG
With the change, we reuse MediaSource's MP4 parsing code to avoid parsing everything twice.
