+++ lib/media/streaming_engine.js
@@ -27,7 +27,6 @@ goog.require('shaka.util.Destroyer');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.FakeEvent');
 goog.require('shaka.util.IDestroyable');
-goog.require('shaka.util.Id3Utils');
 goog.require('shaka.util.LanguageUtils');
 goog.require('shaka.util.ManifestParserUtils');
 goog.require('shaka.util.MimeUtils');
@@ -2250,74 +2249,64 @@ shaka.media.StreamingEngine = class {
     const hasClosedCaptions = stream.closedCaptions &&
         stream.closedCaptions.size > 0;
 
-    let parser;
-    const hasEmsg = ((stream.emsgSchemeIdUris != null &&
-      stream.emsgSchemeIdUris.length > 0) ||
-      this.config_.dispatchAllEmsgBoxes);
+    if (this.config_.shouldFixTimestampOffset) {
+      let parser;
 
-    const isMP4 = stream.mimeType == 'video/mp4' ||
-            stream.mimeType == 'audio/mp4';
-    let timescale = null;
-    if (reference.initSegmentReference) {
-      timescale = reference.initSegmentReference.timescale;
-    }
-    const shouldFixTimestampOffset = isMP4 && timescale &&
-        stream.type === shaka.util.ManifestParserUtils.ContentType.VIDEO &&
-        this.manifest_.type == shaka.media.ManifestParser.DASH &&
-        this.config_.shouldFixTimestampOffset;
+      const isMP4 = stream.mimeType == 'video/mp4' ||
+              stream.mimeType == 'audio/mp4';
+      let timescale = null;
+      if (reference.initSegmentReference) {
+        timescale = reference.initSegmentReference.timescale;
+      }
+      const shouldFixTimestampOffset = isMP4 && timescale &&
+          stream.type === shaka.util.ManifestParserUtils.ContentType.VIDEO &&
+          this.manifest_.type == shaka.media.ManifestParser.DASH;
 
-    if (hasEmsg || shouldFixTimestampOffset) {
-      parser = new shaka.util.Mp4Parser();
-    }
+      if (shouldFixTimestampOffset) {
+        parser = new shaka.util.Mp4Parser();
+      }
 
-    if (hasEmsg) {
-      parser
-          .fullBox(
-              'emsg',
-              (box) => this.parseEMSG_(
-                  reference, stream.emsgSchemeIdUris, box));
-    }
+      if (shouldFixTimestampOffset) {
+        parser
+            .box('moof', shaka.util.Mp4Parser.children)
+            .box('traf', shaka.util.Mp4Parser.children)
+            .fullBox('tfdt', async (box) => {
+              goog.asserts.assert(
+                  box.version != null,
+                  'TFDT is a full box and should have a valid version.');
 
-    if (shouldFixTimestampOffset) {
-      parser
-          .box('moof', shaka.util.Mp4Parser.children)
-          .box('traf', shaka.util.Mp4Parser.children)
-          .fullBox('tfdt', async (box) => {
-            goog.asserts.assert(
-                box.version != null,
-                'TFDT is a full box and should have a valid version.');
+              const parsedTFDT = shaka.util.Mp4BoxParsers.parseTFDTInaccurate(
+                  box.reader, box.version);
 
-            const parsedTFDT = shaka.util.Mp4BoxParsers.parseTFDTInaccurate(
-                box.reader, box.version);
+              const baseMediaDecodeTime = parsedTFDT.baseMediaDecodeTime;
 
-            const baseMediaDecodeTime = parsedTFDT.baseMediaDecodeTime;
-
-            // In case the time is 0, it is not updated
-            if (!baseMediaDecodeTime) {
-              return;
-            }
-            goog.asserts.assert(typeof(timescale) == 'number',
-                'Should be an number!');
-
-            const scaledMediaDecodeTime = -baseMediaDecodeTime / timescale;
-
-            const comparison1 = Number(mediaState.lastTimestampOffset) || 0;
-
-            if (comparison1 < scaledMediaDecodeTime) {
-              const lastAppendWindowStart = mediaState.lastAppendWindowStart;
-              const lastAppendWindowEnd = mediaState.lastAppendWindowEnd;
-              goog.asserts.assert(typeof(lastAppendWindowStart) == 'number',
+              // In case the time is 0, it is not updated
+              if (!baseMediaDecodeTime) {
+                return;
+              }
+              goog.asserts.assert(typeof(timescale) == 'number',
                   'Should be an number!');
-              goog.asserts.assert(typeof(lastAppendWindowEnd) == 'number',
-                  'Should be an number!');
-              await this.setProperties_(mediaState, scaledMediaDecodeTime,
-                  lastAppendWindowStart, lastAppendWindowEnd, reference);
-            }
-          });
-    }
 
-    if (hasEmsg || shouldFixTimestampOffset) {
-      parser.parse(segment, /* partialOkay= */ false, isChunkedData);
+              const scaledMediaDecodeTime = -baseMediaDecodeTime / timescale;
+
+              const comparison1 = Number(mediaState.lastTimestampOffset) || 0;
+
+              if (comparison1 < scaledMediaDecodeTime) {
+                const lastAppendWindowStart = mediaState.lastAppendWindowStart;
+                const lastAppendWindowEnd = mediaState.lastAppendWindowEnd;
+                goog.asserts.assert(typeof(lastAppendWindowStart) == 'number',
+                    'Should be an number!');
+                goog.asserts.assert(typeof(lastAppendWindowEnd) == 'number',
+                    'Should be an number!');
+                await this.setProperties_(mediaState, scaledMediaDecodeTime,
+                    lastAppendWindowStart, lastAppendWindowEnd, reference);
+              }
+            });
+      }
+
+      if (shouldFixTimestampOffset) {
+        parser.parse(segment, /* partialOkay= */ false, isChunkedData);
+      }
     }
 
     await this.evict_(mediaState, presentationTime);
@@ -2340,130 +2329,6 @@ shaka.media.StreamingEngine = class {
         isChunkedData);
     this.destroyer_.ensureNotDestroyed();
     shaka.log.v2(logPrefix, 'appended media segment');
-  }
-
-
-  /**
-   * Parse the EMSG box from a MP4 container.
-   *
-   * @param {!shaka.media.SegmentReference} reference
-   * @param {?Array.<string>} emsgSchemeIdUris Array of emsg
-   *     scheme_id_uri for which emsg boxes should be parsed.
-   * @param {!shaka.extern.ParsedBox} box
-   * @private
-   * https://dashif-documents.azurewebsites.net/Events/master/event.html#emsg-format
-   * aligned(8) class DASHEventMessageBox
-   *    extends FullBox(‘emsg’, version, flags = 0){
-   * if (version==0) {
-   *   string scheme_id_uri;
-   *   string value;
-   *   unsigned int(32) timescale;
-   *   unsigned int(32) presentation_time_delta;
-   *   unsigned int(32) event_duration;
-   *   unsigned int(32) id;
-   * } else if (version==1) {
-   *   unsigned int(32) timescale;
-   *   unsigned int(64) presentation_time;
-   *   unsigned int(32) event_duration;
-   *   unsigned int(32) id;
-   *   string scheme_id_uri;
-   *   string value;
-   * }
-   * unsigned int(8) message_data[];
-   */
-  parseEMSG_(reference, emsgSchemeIdUris, box) {
-    let timescale;
-    let id;
-    let eventDuration;
-    let schemeId;
-    let startTime;
-    let presentationTimeDelta;
-    let value;
-
-    if (box.version === 0) {
-      schemeId = box.reader.readTerminatedString();
-      value = box.reader.readTerminatedString();
-      timescale = box.reader.readUint32();
-      presentationTimeDelta = box.reader.readUint32();
-      eventDuration = box.reader.readUint32();
-      id = box.reader.readUint32();
-      startTime = reference.startTime + (presentationTimeDelta / timescale);
-    } else {
-      timescale = box.reader.readUint32();
-      const pts = box.reader.readUint64();
-      startTime = (pts / timescale) + reference.timestampOffset;
-      presentationTimeDelta = startTime - reference.startTime;
-      eventDuration = box.reader.readUint32();
-      id = box.reader.readUint32();
-      schemeId = box.reader.readTerminatedString();
-      value = box.reader.readTerminatedString();
-    }
-    const messageData = box.reader.readBytes(
-        box.reader.getLength() - box.reader.getPosition());
-
-    // See DASH sec. 5.10.3.3.1
-    // If a DASH client detects an event message box with a scheme that is not
-    // defined in MPD, the client is expected to ignore it.
-    if ((emsgSchemeIdUris && emsgSchemeIdUris.includes(schemeId)) ||
-        this.config_.dispatchAllEmsgBoxes) {
-      // See DASH sec. 5.10.4.1
-      // A special scheme in DASH used to signal manifest updates.
-      if (schemeId == 'urn:mpeg:dash:event:2012') {
-        this.playerInterface_.onManifestUpdate();
-      } else {
-        // All other schemes are dispatched as a general 'emsg' event.
-        const endTime = startTime + (eventDuration / timescale);
-        /** @type {shaka.extern.EmsgInfo} */
-        const emsg = {
-          startTime: startTime,
-          endTime: endTime,
-          schemeIdUri: schemeId,
-          value: value,
-          timescale: timescale,
-          presentationTimeDelta: presentationTimeDelta,
-          eventDuration: eventDuration,
-          id: id,
-          messageData: messageData,
-        };
-
-        // Dispatch an event to notify the application about the emsg box.
-        const eventName = shaka.util.FakeEvent.EventName.Emsg;
-        const data = (new Map()).set('detail', emsg);
-        const event = new shaka.util.FakeEvent(eventName, data);
-        // A user can call preventDefault() on a cancelable event.
-        event.cancelable = true;
-
-        this.playerInterface_.onEvent(event);
-
-        if (event.defaultPrevented) {
-          // If the caller uses preventDefault() on the 'emsg' event, don't
-          // process any further, and don't generate an ID3 'metadata' event
-          // for the same data.
-          return;
-        }
-
-        // Additionally, ID3 events generate a 'metadata' event.  This is a
-        // pre-parsed version of the metadata blob already dispatched in the
-        // 'emsg' event.
-        if (schemeId == 'https://aomedia.org/emsg/ID3' ||
-            schemeId == 'https://developer.apple.com/streaming/emsg-id3') {
-          // See https://aomediacodec.github.io/id3-emsg/
-          const frames = shaka.util.Id3Utils.getID3Frames(messageData);
-          if (frames.length) {
-            /** @private {shaka.extern.ID3Metadata} */
-            const metadata = {
-              cueTime: startTime,
-              data: messageData,
-              frames: frames,
-              dts: startTime,
-              pts: startTime,
-            };
-            this.playerInterface_.onMetadata(
-                [metadata], /* offset= */ 0, endTime);
-          }
-        }
-      }
-    }
   }
 
   /**
@@ -2974,13 +2839,11 @@ shaka.media.StreamingEngine = class {
  *   netEngine: shaka.net.NetworkingEngine,
  *   onError: function(!shaka.util.Error),
  *   onEvent: function(!Event),
- *   onManifestUpdate: function(),
  *   onSegmentAppended: function(!shaka.media.SegmentReference,
  *     !shaka.extern.Stream),
  *   onInitSegmentAppended: function(!number,!shaka.media.InitSegmentReference),
  *   beforeAppendSegment: function(
  *     shaka.util.ManifestParserUtils.ContentType,!BufferSource):Promise,
- *   onMetadata: !function(!Array.<shaka.extern.ID3Metadata>, number, ?number),
  *   disableStream: function(!shaka.extern.Stream, number):boolean
  * }}
  *
@@ -3001,8 +2864,6 @@ shaka.media.StreamingEngine = class {
  *   StreamingEngine.switch*() or StreamingEngine.seeked() to attempt recovery.
  * @property {function(!Event)} onEvent
  *   Called when an event occurs that should be sent to the app.
- * @property {function()} onManifestUpdate
- *   Called when an embedded 'emsg' box should trigger a manifest update.
  * @property {function(!shaka.media.SegmentReference,
  *     !shaka.extern.Stream)} onSegmentAppended
  *   Called after a segment is successfully appended to a MediaSource.
@@ -3012,9 +2873,6 @@ shaka.media.StreamingEngine = class {
  * @property {!function(shaka.util.ManifestParserUtils.ContentType,
  *   !BufferSource):Promise} beforeAppendSegment
  *   A function called just before appending to the source buffer.
- * @property
- *  {!function(!Array.<shaka.extern.ID3Metadata>, number, ?number)} onMetadata
- *   Called when an ID3 is found in a EMSG.
  * @property {function(!shaka.extern.Stream, number):boolean} disableStream
  *   Called to temporarily disable a stream i.e. disabling all variant
  *   containing said stream.
