+++ lib/player.js
@@ -2429,6 +2429,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
             this.processTimedMetadataMediaSrc_(metadata, offset, endTime);
           },
           onEvent: (event) => this.dispatchEvent(event),
+          onManifestUpdate: () => this.onManifestUpdate_(),
         },
         this.lcevcDec_);
     mediaSourceEngine.configure(this.config_.mediaSource);
@@ -3767,7 +3768,6 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       netEngine: this.networkingEngine_,
       onError: (error) => this.onError_(error),
       onEvent: (event) => this.dispatchEvent(event),
-      onManifestUpdate: () => this.onManifestUpdate_(),
       onSegmentAppended: (reference, stream) => {
         this.onSegmentAppended_(
             reference.startTime, reference.endTime, stream.type,
@@ -3781,9 +3781,6 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       },
       beforeAppendSegment: (contentType, segment) => {
         return this.drmEngine_.parseInbandPssh(contentType, segment);
-      },
-      onMetadata: (metadata, offset, endTime) => {
-        this.processTimedMetadataMediaSrc_(metadata, offset, endTime);
       },
       disableStream: (stream, time) => this.disableStream(stream, time),
     };
@@ -4012,6 +4009,16 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
           'manifest.dash.enableAudioGroups configuration',
           'It is now enabled by default and cannot be disabled.');
       delete config['manifest']['dash']['enableAudioGroups'];
+    }
+
+    // Deprecate 'streaming.dispatchAllEmsgBoxes' configuration.
+    if (config['streaming'] && 'dispatchAllEmsgBoxes' in config['streaming']) {
+      shaka.Deprecate.deprecateFeature(5,
+          'streaming.dispatchAllEmsgBoxes configuration',
+          'Please Use mediaSource.dispatchAllEmsgBoxes instead.');
+      config['mediaSource']['dispatchAllEmsgBoxes'] =
+          config['streaming']['dispatchAllEmsgBoxes'];
+      delete config['streaming']['dispatchAllEmsgBoxes'];
     }
 
     // If lowLatencyMode is enabled, and inaccurateManifestTolerance and
