+++ lib/util/player_configuration.js
@@ -235,7 +235,6 @@ shaka.util.PlayerConfiguration = class {
       minBytesForProgressEvents: minBytes,
       preferNativeHls: false,
       updateIntervalSeconds: 1,
-      dispatchAllEmsgBoxes: false,
       observeQualityChanges: false,
       maxDisabledTime: 30,
       // When low latency streaming is enabled, segmentPrefetchLimit will
@@ -402,6 +401,7 @@ shaka.util.PlayerConfiguration = class {
             [cue, uri],
             undefined);
       },
+      dispatchAllEmsgBoxes: false,
     };
 
     let customPlayheadTracker = false;
